refactor: 디렉토리 구조 변경
황지원이 작업한 내역과 기존 atoms 폴더 내에 있던 파일들을 옮겼습니다.
모달은 commons/modals 내부에 위치
