+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import { RecoilRoot } from 'recoil';
 import { Routes, Route, BrowserRouter } from 'react-router-dom';
-import DetailPetPage from 'pages/DetailPetPage';
+import DetailPetPage from 'pages/detailPet/DetailPetPage';
 import Home from 'pages/Home';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import ProfileListPage from 'pages/ProfileListPage';
@@ -10,7 +10,7 @@ import UrgentListPage from 'pages/UrgentListPage';
 import NewListPage from 'pages/NewListPage';
 import LoginPage from 'pages/LoginPage';
 import SignupPage from 'pages/SignupPage';
-import RegisterPage from 'pages/RegisterPage';
+import RegisterPage from 'pages/register/RegisterPage';
 import MapPage from 'pages/MapPage';
 
 const queryClient = new QueryClient();
