+++ src/pages/register/ImageVideoInput.tsx
